fix: Resolve issues related to the file path when importing JSON and separate the basic file operation methods into a dedicated module.
