+++ copy.cjs
@@ -12,3 +12,5 @@ fs.copySync(
 fs.copySync('src/commands/dev/mock', 'dist/commands/dev/config/mock');
 fs.copySync('src/cliconfig.toml', 'dist/cliconfig.toml');
 fs.copySync('src/utils/install', 'dist/utils/install');
+fs.copySync('src/i18n/locales.json', 'dist/i18n/locales.json');
+console.log('copy success');
