+++ src/i18n/index.ts
@@ -3,8 +3,11 @@ import { resolve } from 'path';
 import { getCliConfig } from '../utils/fileUtils/index.js';
 import logger from '../libs/logger.js';
 import { readJson } from '../utils/readJson.js';
+import { getDirName } from '../utils/fileUtils/base.js';
 
-const localesJson = readJson(resolve('./locales.json'));
+const localesJson = readJson(
+  resolve(getDirName(import.meta.url), './locales.json')
+);
 
 interface Locales {
   [key: string]: {
